refactor(commands): new commands folder structure
