+++ commands/all.go
@@ -1,10 +1,18 @@
 package commands
 
 import (
-	_ "github.com/versenilvis/iris/commands/dev"
+	_ "github.com/versenilvis/iris/commands/cc"
 	_ "github.com/versenilvis/iris/commands/fs"
+	_ "github.com/versenilvis/iris/commands/git"
+	_ "github.com/versenilvis/iris/commands/golang"
+	_ "github.com/versenilvis/iris/commands/js"
+	_ "github.com/versenilvis/iris/commands/jvm"
 	_ "github.com/versenilvis/iris/commands/ops"
+	_ "github.com/versenilvis/iris/commands/pkginstaller"
+	_ "github.com/versenilvis/iris/commands/python"
 	_ "github.com/versenilvis/iris/commands/runner"
+	_ "github.com/versenilvis/iris/commands/rust"
 	_ "github.com/versenilvis/iris/commands/sys"
 	_ "github.com/versenilvis/iris/commands/text"
+	_ "github.com/versenilvis/iris/commands/view"
 )
