+++ commands/python/pip.go
@@ -1,4 +1,4 @@
-package dev
+package python
 
 import (
 	"context"
