+++ commands/python/python.go
@@ -1,4 +1,4 @@
-package dev
+package python
 
 import (
 	"github.com/versenilvis/iris/commands/core"
