+++ commands/rust/cargo.go
@@ -1,4 +1,4 @@
-package dev
+package rust
 
 import (
 	"github.com/versenilvis/iris/commands/core"
